Ensure encoded base64 is URL safe
Crypto.JS (as pointed out by @cfanboy: https://github.com/kartenmacherei/expo-jwt/issues/1#issuecomment-417581629)
encodes Base 64 that is not URL safe. This was preventing JWTs encoded
by this library to fail on the validator of https://jwt.io/

This commit adds a helper that converts the Crypto.js base64 to a URL
safe version.