Support all headers from JWT 7515 Spec
Support all new optional header claims from the 7515 spec:
https://www.rfc-editor.org/rfc/rfc7515#section-4.1

Allow these to be set using the existing API.

Additionally allows the API to accept simply `alg` instead of the custom
`algorithm` option as some suers may expect.

The `buildHeader` public method probably should never have been there
and now simply returns the `this.header` set in the constructor, but I
have left it there with a comment, for backwards compatibility.

In future this can be removed if we bump the major version in order to
clean up the API.